Prevent having a null value on from and to
The reason why to prevent this from is because this will give an error. This doesn't actually make sense for gdscript programmer because you don't actually had to put null on the value, but on those who is making a wrapper on other language to use the addon, this gets annoying, especially when working with relative values.